Print the output of the subprocess.

+++ update.py
@@ -336,7 +336,7 @@ def _parse_args():
                                 directory (default: %(default)s))""")
     return parser.parse_args()
 
-if __name__ == "__main__":
+def update():
     args = _parse_args()
     tests = args.tests
     target = args.target
@@ -363,10 +363,21 @@ if __name__ == "__main__":
     create_test_lists(tests, target, ctx)
     hg_target = os.path.join(target, DFL_REPO)
     print "cloning %s, this can take some minutes" % args.dfl_repo
-    cmd_call("hg", "clone", args.dfl_repo, hg_target)
+    out, err = cmd_call("hg", "clone", args.dfl_repo, hg_target)
+    if err:
+        print err
+        return
+    print out
     print "updating the repo to %s" % args.revision
-    cmd_call("hg", "up", args.revision)
+    out, err = cmd_call("hg", "up", args.revision)
+    if err:
+        print err
+        return
+    print out
     print "copying %s to %s" % (os.path.join(hg_target, "src"),
                                 os.path.join(target, DFLSRC))
     shutil.copytree(os.path.join(hg_target, "src"), os.path.join(target, DFLSRC))
     shutil.rmtree(hg_target)
+
+if __name__ == "__main__":
+    update()
